+++ R/sort_omv.R
@@ -3,6 +3,7 @@
 #' @param fleInp Name (including the path, if required) of the data file to be read ("FILENAME.ext"; default: ""); can be any supported file type, see Details below
 #' @param fleOut Name (including the path, if required) of the data file to be written ("FILENAME.omv"; default: ""); if empty, the extension of fleInp is replaced with "_sorted(file extension -> .omv)"
 #' @param varSrt Variable(s) that are used to sort the data frame (see Details; default: c())
+#' @param psvAnl Whether analyses that are contained in the input file shall be transferred to the output file (TRUE / FALSE; default: FALSE)
 #' @param usePkg Name of the package: "foreign" or "haven" that shall be used to read SPSS, Stata and SAS files; "foreign" is the default (it comes with base R), but "haven" is newer and more comprehensive
 #' @param selSet Name of the data set that is to be selected from the workspace (only applies when reading .RData-files)
 #' @param ... Additional arguments passed on to methods; see Details below
@@ -42,7 +43,7 @@
 #'
 #' @export sort_omv
 #'
-sort_omv <- function(fleInp = c(), fleOut = "", varSrt = c(), usePkg = c("foreign", "haven"), selSet = "", ...) {
+sort_omv <- function(fleInp = c(), fleOut = "", varSrt = c(), psvAnl = FALSE, usePkg = c("foreign", "haven"), selSet = "", ...) {
     if (length(varSrt) == 0 || !all(nzchar(varSrt))) {
         stop("Calling sort_omv requires giving at least one variable to sort after.")
     }
@@ -59,6 +60,9 @@ sort_omv <- function(fleInp = c(), fleOut = "", varSrt = c(), usePkg = c("foreig
 
     # write file
     write_omv(dtaFrm, fleOut)
+
+    # transfer analyses from input to output file
+    if (psvAnl) xfrAnl(fleInp, fleOut)
 }
 
 srtFrm <- function(dtaFrm = NULL, varSrt = c()) {