+++ R/write_omv.R
@@ -36,7 +36,7 @@
 #' print(list.files(dirname(fleOMV), basename(fleOMV)))
 #' # -> "file[...].omv" ([...] is a combination of random numbers / characters
 #' print(file.info(fleOMV)$size)
-#' # -> 2448 (size may differ on different OSes)
+#' # -> approx. 2500 (size may differ on different OSes)
 #' unlink(fleOMV)
 #' }
 #'